only delete directory if it exists
git-svn-id: https://svn.kwarc.info/repos/MMT/src@2798 09069c07-60fa-4895-8316-727242e252e4

+++ jEdit-mmt/src/info/kwarc/mmt/jeditsetup/Setup.scala
@@ -97,8 +97,10 @@ object Setup {
     }
     if (!install) {
       val d = jedit / plug
-      d.deleteDir
-      println("deleting " + d)
+      if (d.isDirectory) {
+        d.deleteDir
+        println("deleting directory " + d)
+      }
     }
   }
